Adds support for test and adds first for example
